+++ changelog.md
@@ -2,11 +2,13 @@
 
 ## HEAD
 
+## 6.1 2026-07-27 [Commit 3b8d738](https://github.com/allen-garvey/dithermark/commit/3b8d7388a41ac3467ca58b897671562caad68922)
+
 ### Features
 
 * Add R2 Sequence noise dither and R2 Sequence ordered dither variants
 
-* Add ability to set video quality when exporting video.
+* Add the ability to set video quality when exporting video.
 
 ### Technical